+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+- Add new [EnvironmentActivator](docs/activator/environment.md) @migo315
+
 ## [1.0.0]
 ### Added
 - Initial commit and relase @migo315